Adding user system model and authentication.
Adding internationalization for english and spanish.

+++ database/migrations/2025_10_12_000004_create_course_enrollments_table.php
@@ -1,42 +1,45 @@
 <?php
 
+use Database\Common\DatabaseConnections as DB_CONN;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
 return new class() extends Migration
 {
-    protected $connection = 'operational';
+    protected $connection = DB_CONN::OPERATIONAL;
 
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::connection('operational')->create('course_enrollments', function (Blueprint $table) {
-            $table->uuid('id')->primary();
-            $table->uuid('tenant_id');
-            $table->uuid('user_id');
-            $table->uuid('course_id');
-            $table->date('enrolled_at');
-            $table->boolean('is_completed')->default(false);
-            $table->timestamps();
+        if (! Schema::connection($this->connection)->hasTable('course_enrollments')) {
+            Schema::connection($this->connection)->create('course_enrollments', function (Blueprint $table) {
+                $table->uuid('id')->primary();
+                $table->uuid('tenant_id');
+                $table->uuid('user_id');
+                $table->uuid('course_id');
+                $table->date('enrolled_at');
+                $table->boolean('is_completed')->default(false);
+                $table->timestamps();
 
-            $table->foreign('tenant_id')
-                ->references('id')
-                ->on('tenants')
-                ->onDelete('cascade');
+                $table->foreign('tenant_id')
+                    ->references('id')
+                    ->on('tenants')
+                    ->onDelete('cascade');
 
-            $table->foreign('user_id')
-                ->references('id')
-                ->on('users')
-                ->onDelete('cascade');
+                $table->foreign('user_id')
+                    ->references('id')
+                    ->on('users')
+                    ->onDelete('cascade');
 
-            $table->foreign('course_id')
-                ->references('id')
-                ->on('courses')
-                ->onDelete('cascade');
-        });
+                $table->foreign('course_id')
+                    ->references('id')
+                    ->on('courses')
+                    ->onDelete('cascade');
+            });
+        }
     }
 
     /**
@@ -44,6 +47,6 @@ return new class() extends Migration
      */
     public function down(): void
     {
-        Schema::connection('operational')->dropIfExists('course_enrollments');
+        Schema::connection($this->connection)->dropIfExists('course_enrollments');
     }
 };
